requirements.txt
# pip install -r requirements.txt
openai
python-dotenv